Ca avance, mais pas comme il faut

+++ core/src/com/gdx/game/model/GameElement.java
@@ -20,11 +20,11 @@ public abstract class GameElement {
         this.position = position;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.position.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.position.y = y;
     }
